content pipeline- only
this is the test for the content pipeline . the added files are: Level
Library
This way we can input external files in the project other than the
official.
Right now the only thing it can do is implement a level using a matrix
with ints.
Its not fully operational yet . I need the actual code that loads
everything, and draws the tiles.. Now we can change it in the future to
be able to read xml files.
I'm the best!!!
Content Type Reader
Mylevels
Content Processor
Level extrension Lib
Lvl Writer kai Reader